+++ crates/compiler/src/driver_constants.rs
@@ -223,6 +223,8 @@ pub const ECHO_BUFFER_MAX_EDL: u8 = 15;
 pub const MAX_SUBROUTINES: usize = 255;
 pub const MAX_SFX_SUBROUTINES: u8 = 255;
 
+pub(crate) const BLANK_SONG_BIN: [u8; 1] = [0];
+
 pub const SONG_HEADER_SIZE: usize = 16;
 pub const SONG_HEADER_ACTIVE_MUSIC_CHANNELS: usize = 0;
 pub const SONG_HEADER_ECHO: usize = 1;